requirements.txt
gunicorn==20.1.0
flask==2.0.1
requests
google-cloud-secret-manager
psycopg2
sqlalchemy
python-dateutil
pytz
openai>=0.27.0
gmplot
polyline
jinja2